move absorbance plate block to softmax pro structure
